restructure scrapers
