+++ scrapers/covid19/getWorldometers.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const cheerio = require('cheerio');
-const countryUtils = require('../utils/countryUtils');
-const logger = require('../utils/logger');
+const countryUtils = require('../../utils/countryUtils');
+const logger = require('../../utils/logger');
 
 const columns = ['index', 'country', 'cases', 'todayCases', 'deaths', 'todayDeaths', 'recovered', 'todayRecovered', 'active',
 	'critical', 'casesPerOneMillion', 'deathsPerOneMillion', 'tests', 'testsPerOneMillion', 'population', 'continent', 'oneCasePerPeople', 'oneDeathPerPeople', 'oneTestPerPeople'];
