+++ scrapers/covid19/govScrapers/getItaly.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
-const logger = require('../../utils/logger');
-const csvUtils = require('../../utils/csvUtils');
+const logger = require('../../../utils/logger');
+const csvUtils = require('../../../utils/csvUtils');
 
 const url = 'https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-regioni/dpc-covid19-ita-regioni-latest.csv';
