+++ scrapers/covid19/govScrapers/getNigeria.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const cheerio = require('cheerio');
-const logger = require('../../utils/logger');
+const logger = require('../../../utils/logger');
 
 const columns = ['state', 'cases', 'active', 'recovered', 'deaths'];
 /**
