+++ scrapers/covid19/govScrapers/getSwitzerland.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const csv = require('csvtojson');
-const logger = require('../../utils/logger');
+const logger = require('../../../utils/logger');
 
 /**
  * Scrapes Swiss government site and fills array of data from table
