+++ scrapers/covid19/jhuLocations.js
@@ -1,6 +1,6 @@
 const axios = require('axios');
 const csv = require('csvtojson');
-const logger = require('../utils/logger');
+const logger = require('../../utils/logger');
 const base = 'https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports/';
 
 /**
